requirements.txt
requests>=2.31.0
watchdog>=3.0.0
Pillow>=10.0.0
beautifulsoup4>=4.12.0
PyYAML>=6.0.0
pytest>=7.4.0
flake8>=6.0.0
black>=23.0.0
